All simulation classes + process class + abstract scheduling class  check the description + a few adjustment in GUI
all the Gui has been completed we just need to finish all the simulation algorithms and show it using jfreechart.

scheduling is an abstract class in which all the simulation algorithms gonna inherit from.

the process class holds all the process data and when we add a new process in the table we initiate a new process and add it to the vector p (and pass it as a parameter to the process_simulation_panel ).